// 04_03-04--main/04_03-04--main/src/features/comments/types/comments/output.ts
import { LikeStatus } from '../../../posts/types/likes/input';

export type OutputCommentType = {
  id: string;
  content: string;
  commentatorInfo: {
    userId: string;
    userLogin: string;
  };
  createdAt: string;
  likesInfo: {
    likesCount: number;
    dislikesCount: number;
    myStatus: LikeStatus;
  };
};